chore: move r6-runner.mjs to scripts/openrouter-live-check.mjs + handle missing .env
The opt-in OpenRouter live-check runner sat at the repo root under an opaque
name. Move it to scripts/openrouter-live-check.mjs (matching its npm script) and:

- Catch ENOENT/permission errors from reading .env and exit 2 with a short,
  user-friendly message instead of an unhandled exception + stack trace
  (Copilot review).
- Resolve .env and the vitest cwd relative to the repo root (the script now
  lives one level down), and pin spawnSync cwd so it runs from any directory.

package.json `openrouter-live-check` updated to the new path. No other reference
to the old filename exists.
